+++ python/marshmallow_recipe/json_schema.py
@@ -194,6 +194,12 @@ def __convert_field_to_json_schema(
 
     if field_type is str:
         schema["type"] = "string"
+        if metadata.get("min_length") is not None:
+            schema["minLength"] = metadata["min_length"]
+        if metadata.get("max_length") is not None:
+            schema["maxLength"] = metadata["max_length"]
+        if metadata.get("regexp") is not None:
+            schema["pattern"] = metadata["regexp"]
     elif field_type is int:
         schema["type"] = "integer"
     elif field_type is float: